example/client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description =  "HTTP Streamable based example MCP Client"
requires-python = ">=3.12"
dependencies = [
    "anthropic~=0.51.0",
    "mcp~=1.9.0",
    "python-dotenv~=1.1.0",
    "google-genai~=1.27.0",
]